renamed hdfmappers.py to hdfmapper.py
